Migrate code to main files
