--- pkg/presentation/rest/tests.go
@@ -1,19 +1,0 @@
-package rest
-
-import (
-	"net/http"
-	"net/http/httptest"
-	"testing"
-)
-
-func executeRequest(req *http.Request) *httptest.ResponseRecorder {
-	rw := httptest.NewRecorder()
-	router.ServeHTTP(rw, req)
-	return rw
-}
-
-func checkResponseCode(t *testing.T, expected, actual int) {
-	if expected != actual {
-		t.Errorf("Expected response code %d. Got %d\n", expected, actual)
-	}
-}
